// src/app/modules/admission/models/student.admission.info.ts
export class StudentAdmissionInfo {
    STUDENT_ID: number
    ORG_CODE: string
    ORG_NAME: string
    STUDENT_CODE: string
    STUDENT_NAME: string
    STUDENT_IMAGE: string
    ADMISSION_DATE: string
    START_SESSION_CODE: string
    SESSION_NAME: string
    CUR_SESSION_CODE: string
    SESSION_NAME_1: string
    CAMPUS_CODE: string
    CAMPUS_NAME: string
    CLASS_CODE: string
    CLASS_NAME: string
    GROUP_CODE: string
    GROUP_NAME: string
    SECTION_CODE: string
    SECTION_NAME: string
    ADMISSION_TEST_MARK: string
    ADMISSION_TEST_POS: string
    CLASS_ROLL: string
    GENDER_CODE: string
    GENDER_NAME: string
    SMS_MOBILE_NUM: string
    DATE_OF_BIRTH: string
    BOARD_UNIV_CODE: string
    BOARD_UNIV_NAME: string
    SHIFT_CODE: string
    SHIFT_NAME: string
    VERSION_CODE: string
    VERSION_NAME: string
    RELIGION_CODE: string
    RELIGION_NAME: string
    BLOOD_GROUP_CODE: string
    BLOOD_GROUP_NAME: string
    NATIONALITY_CODE: string
    NATIONALITY_NAME: string
    BIRTH_PLACE_CODE: string
    BIRTH_PLACE: string
    LAST_SCHOOL_NAME_ADDR: string
    PRESENT_ADDR: string
    PERMANENT_ADDR: string
    FATHERS_NAME: string
    FATHER_OCCUPATION: string
    OCCUPATION_NAME: string
    FATHER_CONTACT_NO: string
    FATHERS_IMAGE: string
    MOTHERS_NAME: string
    MOTHER_OCCUPATION: string
    OCCUPATION_NAME_1: string
    MOTHER_CONTACT_NO: string
    MOTHERS_IMAGE: string
    GURDIAN_NAME: string
    GUARDIAN_OCCUPATION: string
    OCCUPATION_NAME_2: string
    GUARDIAN_CONTACT_NO: string
    YEAR_CODE: string
    SEMESTER_CODE: string
    GUARDIAN_IMAGE: string
    GUARDIAN_RELATION_TYPE_CODE: string
    RELATION_TYPE_NAME: string
    REMARKS: string
    ACTIVE_STATUS: string
    USER_NAME:string
    RowStatus:number
  }
  

  export class StudentAdmissionForm {
    STUDENT_ID: number
    ORG_CODE: string
    ORG_NAME: string
    CAMPUS_CODE: string
    CAMPUS_NAME: string
    CLASS_CODE: string
    CLASS_NAME: string
    GROUP_CODE: string
    GROUP_NAME: string
    START_SESSION_CODE: string
    CUR_SESSION_CODE: string
    SECTION_CODE: string
    SECTION_NAME: string
    SHIFT_CODE: string
    SHIFT_NAME: string
    VERSION_CODE: string
    VERSION_NAME: string
    YEAR_CODE: string
    SEMESTER_CODE: string

    CLASS_ROLL: string
    STUDENT_CODE: string
    STUDENT_NAME: string
    SUR_NAME: string
    STUDENT_IMAGE: any;
    STUDENT_IMG: any;
    MyFile: any
    DATE_OF_BIRTH: string
    ADMISSION_DATE: string
    BIRTH_REG_NUMBER: string
    GENDER_CODE: string
    GENDER_NAME: string
    BLOOD_GROUP_CODE: string
    BLOOD_GROUP_NAME: string
    RELIGION_CODE: string
    RELIGION_NAME: string

    FATHERS_NAME: string
    FATHERS_NID: string
    FATHER_OCCUPATION: string
    OCCUPATION_NAME: string
    FATHER_CONTACT_NO: string
    MOTHERS_NAME: string
    MOTHERS_NID: string
    MOTHER_OCCUPATION: string
    OCCUPATION_NAME_1: string
    MOTHER_CONTACT_NO: string
   
    GURDIAN_NAME: string
    GUARDIAN_NID: string
    GUARDIAN_OCCUPATION: string
    OCCUPATION_NAME_2: string
    GUARDIAN_CONTACT_NO: string
    GUARDIAN_RELATION_TYPE_CODE: string
    RELATION_TYPE_NAME: string
   
    SMS_MOBILE_NUM:string;
    PRESENT_ADDR: string
    PERMANENT_ADDR: string
    
    SIBLING_STUDENT_CODE:string;
    SIBLING_CLASS_CODE:string
    FACEBOOK_ID:string
    EMAIL_ADDRESS:string
    TRANSPORT:string;
    USER_NAME:string
    RowStatus:number
  }