// src/app/modules/admission/models/studentapproveparams.ts
export class Studentapproveparams {
    ORG_CODE: string;
    CAMPUS_CODE: string;
    CLASS_CODE: string;
    START_DATE: any;
    END_DATE: any;
    SESSION_CODE: string;
}
export class StudentapproveMst {
    ORG_CODE:string;
    CAMPUS_CODE:string;
    CLASS_CODE: string;
    SESSION_CODE: string;
    USER_CODE: string;
    STUDENT_CODE:any[];
    ADMISSION_ROLL:any[];
    ROW_STATUS:number;

}